PlantSitter/requirements.txt
djangorestframework
drf_yasg
Pillow